requirements.txt
Django>=4.2,<5.0
supabase>=2.0
python-decouple>=3.8
whitenoise>=6.5
gunicorn>=21.0
djangorestframework>=3.14
django-cors-headers>=4.3

